Add an options object to turn on/off debugging
The console.logs in the plugin can be very useful when setting up gulp
tasks, but they're a huge annoyance after that point. This change allows
these logs to be turned on or off with a settings object that can also
be used for any other customization options needed in the future.